Update and rename nuage_overcloud_full_patch_13.py to nuage_overcloud_full_patch.py
